% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_analyze.R
\name{clean_data}
\alias{clean_data}
\title{Clean Data Function}
\usage{
clean_data(data)
}
\arguments{
\item{data}{A data frame containing the data to be cleaned.}
}
\value{
A cleaned data frame.
}
\description{
This function cleans the input data by removing missing values, removing duplicates, and converting column names to lowercase.
}
\examples{
data <- data.frame(A = c(1, 2, NA, 4, 4), B = c("a", "b", "c", "d", "d"))
clean_data(data)
}
